requirements.txt
python-dotenv
openai
tenacity
graphviz
streamlit
pyyaml
duckduckgo_search
beautifulsoup4
requests
tavily-python
fastapi